[CHORE/#7] 컴포넌트 파일화
