+++ app/src/main/java/com/sopt/bubble/feature/friends/detail/component/DetailTopBar.kt
@@ -1,0 +1,36 @@
+package com.sopt.bubble.feature.friends.detail.component
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.dp
+import com.sopt.bubble.R
+import com.sopt.bubble.ui.theme.Gray200
+
+@Composable
+fun DetailTopBar(modifier: Modifier) {
+    Row(
+        modifier = modifier
+            .fillMaxWidth()
+            .background(Gray200)
+            .padding(horizontal = 31.dp, vertical = 52.dp),
+        horizontalArrangement = Arrangement.SpaceBetween,
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        Image(
+            painter = painterResource(id = R.drawable.ic_detail_close),
+            contentDescription = null
+        )
+        Image(
+            painter = painterResource(id = R.drawable.ic_detail_empty_star),
+            contentDescription = null
+        )
+    }
+}
